mcp/requirements.txt
mcp
snowflake-connector-python[pandas]
